+++ backend/bin/run-jobs.sh
@@ -6,5 +6,5 @@ ${DIR}/console update-chars --log --hide-details
 ${DIR}/console update-player-groups --log --hide-details
 ${DIR}/console check-tokens --log --hide-details
 ${DIR}/console update-member-tracking --log
-${DIR}/console send-account-disabled-mail --log
+${DIR}/console send-invalid-token-mail --log
 ${DIR}/console clean-http-cache --log