+++ backend/src/Application.php
@@ -17,7 +17,7 @@ use Neucore\Command\DBVerifySSL;
 use Neucore\Command\DoctrineFixturesLoad;
 use Neucore\Command\MakeAdmin;
 use Neucore\Command\RevokeToken;
-use Neucore\Command\SendAccountDisabledMail;
+use Neucore\Command\SendInvalidTokenMail;
 use Neucore\Command\UpdateCharacters;
 use Neucore\Command\UpdateMemberTracking;
 use Neucore\Command\UpdatePlayerGroups;
@@ -396,7 +396,7 @@ class Application
         $console->add($this->container->get(UpdateCharacters::class));
         $console->add($this->container->get(CheckTokens::class));
         $console->add($this->container->get(UpdatePlayerGroups::class));
-        $console->add($this->container->get(SendAccountDisabledMail::class));
+        $console->add($this->container->get(SendInvalidTokenMail::class));
         $console->add($this->container->get(UpdateMemberTracking::class));
         $console->add($this->container->get(DoctrineFixturesLoad::class));
         $console->add($this->container->get(DBVerifySSL::class));