+++ frontend/src/pages/SystemSettings--Mails.vue
@@ -13,27 +13,26 @@
         <span class="small">The character is used for all mails.</span>
     </div>
 
-    <div class="card-header">Notification for deactivated accounts</div>
+    <div class="card-header">Invalid ESI token</div>
     <div class="card-body">
         <p>
-            This EVE mail is sent when an account has been deactivated
-            because one of its characters contains an invalid or no ESI token.<br>
-            The mail is only sent if that feature is activated above.
+            This EVE mail is sent when an account contains a character with an invalid or no ESI token.<br>
+            Accounts with the "managed" status are excluded from this.
         </p>
 
-        <button class="btn btn-success btn-sm" v-on:click="sendMailAccountDisabledTestMail()">Send test mail</button>
+        <button class="btn btn-success btn-sm" v-on:click="sendMailInvalidTokenTestMail()">Send test mail</button>
         <small>Mail will be send to the logged-in user.</small>
 
         <div class="custom-control custom-checkbox mb-2 mt-3">
             <input class="custom-control-input" type="checkbox" value="1"
-                   id="mail_account_disabled_active" name="mail_account_disabled_active"
-                   :checked="settings.mail_account_disabled_active === '1'"
-                   @change="$emit('changeSetting', 'mail_account_disabled_active', $event.target.checked ? '1' : '0')">
-            <label class="custom-control-label" for="mail_account_disabled_active">Activate mail</label>
+                   id="mail_invalid_token_active" name="mail_invalid_token_active"
+                   :checked="settings.mail_invalid_token_active === '1'"
+                   @change="$emit('changeSetting', 'mail_invalid_token_active', $event.target.checked ? '1' : '0')">
+            <label class="custom-control-label" for="mail_invalid_token_active">Activate mail</label>
         </div>
         <div class="form-group">
             <label class="col-form-label">Alliances</label>
-            <multiselect v-model="mailAccountDisabledAlliances" :options="alliances"
+            <multiselect v-model="mailInvalidTokenAlliances" :options="alliances"
                          label="name" track-by="id" :multiple="true"
                          :loading="false" :searchable="true"
                          placeholder="Select alliances"></multiselect>
@@ -44,16 +43,16 @@
             </small>
         </div>
         <div class="form-group">
-            <label class="col-form-label" for="mailAccountDisabledSubject">Subject</label>
-            <input id="mailAccountDisabledSubject" type="text" class="form-control"
-                   v-model="settings.mail_account_disabled_subject"
-                   v-on:input="$emit('changeSettingDelayed', 'mail_account_disabled_subject', $event.target.value)">
+            <label class="col-form-label" for="mailInvalidTokenSubject">Subject</label>
+            <input id="mailInvalidTokenSubject" type="text" class="form-control"
+                   v-model="settings.mail_invalid_token_subject"
+                   v-on:input="$emit('changeSettingDelayed', 'mail_invalid_token_subject', $event.target.value)">
         </div>
         <div class="form-group">
-            <label for="mailAccountDisabledBody">Message</label>
-            <textarea v-model="settings.mail_account_disabled_body" class="form-control"
-                      v-on:input="$emit('changeSettingDelayed', 'mail_account_disabled_body', $event.target.value)"
-                      id="mailAccountDisabledBody" rows="6"></textarea>
+            <label for="mailInvalidTokenBody">Message</label>
+            <textarea v-model="settings.mail_invalid_token_body" class="form-control"
+                      v-on:input="$emit('changeSettingDelayed', 'mail_invalid_token_body', $event.target.value)"
+                      id="mailInvalidTokenBody" rows="6"></textarea>
         </div>
     </div>
 </div>
@@ -72,7 +71,7 @@ export default {
         return {
             alliances: [],
             alliancesLoaded: false,
-            mailAccountDisabledAlliances: null,
+            mailInvalidTokenAlliances: null,
         }
     },
 
@@ -85,7 +84,7 @@ export default {
             readSettings(this);
         },
 
-        mailAccountDisabledAlliances (newValues, oldValues) {
+        mailInvalidTokenAlliances (newValues, oldValues) {
             if (oldValues === null) {
                 return;
             }
@@ -94,13 +93,13 @@ export default {
                 oldAllianceIds.push(oldValue.id);
             }
             const newAllianceIds = [];
-            for (let alliance of this.mailAccountDisabledAlliances) {
+            for (let alliance of this.mailInvalidTokenAlliances) {
                 newAllianceIds.push(alliance.id);
             }
             if (newAllianceIds.join(',') === oldAllianceIds.join(',')) {
                 return;
             }
-            this.$emit('changeSetting', 'mail_account_disabled_alliances', newAllianceIds.join(','));
+            this.$emit('changeSetting', 'mail_invalid_token_alliances', newAllianceIds.join(','));
         },
     },
 
@@ -110,9 +109,9 @@ export default {
             this.$emit('changeSetting', 'mail_character', '');
         },
 
-        sendMailAccountDisabledTestMail () {
+        sendMailInvalidTokenTestMail () {
             const vm = this;
-            new SettingsApi().sendAccountDisabledMail((error, data) => {
+            new SettingsApi().sendInvalidTokenMail((error, data) => {
                 if (error) { // 403 usually
                     return;
                 }
@@ -143,17 +142,17 @@ function readSettings(vm) {
         return; // wait for alliance list and settings
     }
 
-    vm.mailAccountDisabledAlliances = [];
+    vm.mailInvalidTokenAlliances = [];
 
     for (const [name, value] of Object.entries(vm.settings)) {
-        if (name !== 'mail_account_disabled_alliances') {
+        if (name !== 'mail_invalid_token_alliances') {
             continue;
         }
         for (let allianceId of value.split(',')) {
             allianceId = parseInt(allianceId);
             for (let alliance of vm.alliances) {
                 if (alliance.id === allianceId) {
-                    vm.mailAccountDisabledAlliances.push(alliance);
+                    vm.mailInvalidTokenAlliances.push(alliance);
                     break;
                 }
             }